Org and rungroups (#175)
* changing file structure for organization

* changed makefile

* cleaning up rungroup files and organizing structure some

* more cleaning

* deleting commit/query line from two files

---------
