+++ src/api/lambdas/bedrock-api-backend/assets/deleteAsset.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-console */
 const { Client } = require('pg');
-const pgErrorCodes = require('./pgErrorCodes');
+const pgErrorCodes = require('../pgErrorCodes');
 
 async function newClient(connection) {
   const client = new Client(connection);
