+++ src/api/lambdas/bedrock-api-backend/assets/handleAssets.js
